refactor: move safety analysis logic to utils
